1. fixed some interface name 2. updated curl.txt 3. removed "receivers" parameter in sharedata interface

+++ golang/datamgr/src/apiv1/json.go
@@ -83,7 +83,7 @@ type ShareInfoData struct{
 	CrTime  string `json:"crtime"`
 //	FileUri string `json:"fileuri"`
 	OrgName string `json:"orgname"`
-	IsDir int `json:"isdir"`
+	IsDir byte `json:"isdir"`
 }
 
 type IShareInfoAck struct{
